take the selection controls out of the groups and put them in their own widget, for now parked together with the editing controls
